Cookie expiration date now extracted from cookie.
Displays a warning if there are less than 7 days before the cookie expires. This is the variable checked in the healthcheck.
Changed conditions from INFO to ERROR in cases where the script cannot continue.
If script is run with the GENERATECOOKIE=True variable, script now removes cookie prior to generating the new one.