Various - see description
Now handles non-2FA cookies and doesn't hang
If 2FA enabled, cookie is not created and session is non-interactive session, it will now prompt for cookie to be created and keep retrying.
Conversely, if 2FA is enabled, cookie is not created, it will only allow for cookie generation in an interactive session. The makes the GENERATECOOKIE variable redundant.
If cookie exists, automatically detects whether it's 2FA enabled.