Healthcheck not working as expected
Healthcheck wasn't working correctly. A variable was empty so the script was exiting with 'out of range' which, rather oddly, gave exit code 0 instead of a failure. I have also added a exit code check for the icloudpd process so that the container will be marked as unhealthy if the download fails... Incorrect password for example.