Bump version to 1.29.4
As version 1.27.5 does not exist on pypi anymore
and it has been set in 980ba67ccec7a9a02b5fc76970062cb48572cf40
we should try to use latest version available.

I have already built and tested this version on my server
tdudlak/icloudpd:1.29.4 is present on dockerhub
Built was done when tag was not present in repo yet so i used git.
Here it is properly using the pypi release.

Signed-off-by: Tibor Dudlák <tibor.dudlak@gmail.com>